some play with evolution

+++ src/main/java/org/fruttech/neurogene/DataLine.java
@@ -3,6 +3,6 @@ package org.fruttech.neurogene;
 import java.util.List;
 
 public class DataLine {
-    List<Float> input;
-    List<Float> output;
+    List<Double> input;
+    List<Double> output;
 }
